+++ packages/saber/lib/utils/parseAttributes.js
@@ -13,32 +13,40 @@ module.exports = (content, filepath) => {
     plugins: ['typescript', 'jsx', 'objectRestSpread', 'classProperties']
   })
 
-  let attributes = {}
+  let data = {}
 
   traverse.default(ast, {
     ObjectExpression(path) {
-      const isAttributes =
+      const name =
         path.parent &&
         path.parent.type === 'VariableDeclarator' &&
-        path.parent.id.name === 'attributes'
-      const isConstAttributes =
-        isAttributes &&
-        path.parentPath.parent &&
-        path.parentPath.parent.kind === 'const'
-      const isExportedConstAttributes =
-        isConstAttributes &&
+        path.parent.id.name
+
+      if (!['attributes', 'data'].includes(name)) {
+        return
+      }
+
+      const isConst =
+        path.parentPath.parent && path.parentPath.parent.kind === 'const'
+      if (!isConst) {
+        return
+      }
+
+      const isExport =
         path.parentPath.parentPath &&
         path.parentPath.parentPath.parent.type === 'ExportNamedDeclaration'
-      if (isExportedConstAttributes) {
-        const { confident, value } = path.evaluate()
-        if (confident) {
-          attributes = value
-          path.node.properties = []
-        } else {
-          throw new Error(
-            `"attributes" is supposed to have the same value when executed in runtime and build time.`
-          )
-        }
+      if (!isExport) {
+        return
+      }
+
+      const { confident, value } = path.evaluate()
+      if (confident) {
+        data = value
+        path.node.properties = []
+      } else {
+        throw new Error(
+          `"${name}" is supposed to have the same value when executed in runtime and build time.`
+        )
       }
     }
   })
@@ -46,7 +54,7 @@ module.exports = (content, filepath) => {
   const { code } = generator.default(ast)
 
   return {
-    attributes,
+    data,
     code
   }
 }