+++ packages/saber/vue-renderer/app/theme/layouts/default.vue
@@ -1,6 +1,6 @@
 <template>
   <div>
-    <h2 v-if="page.attributes.title">{{ page.attributes.title }}</h2>
+    <h2 v-if="page.title">{{ page.title }}</h2>
     <slot name="default"></slot>
   </div>
 </template>