+++ website/src/components/Leftbar.vue
@@ -1,6 +1,8 @@
 <template>
   <div class="leftbar" :class="{'is-hidden': hide}">
     <div class="sticky">
+      <SiteNav class="show-on-mobile" />
+
       <div class="items">
         <div class="item" v-for="(item, i) in items" :key="i">
           <div class="item-title" @click="toggleOpenLink(item.children[0].link)">{{ item.title }}</div>
@@ -19,7 +21,13 @@
 </template>
 
 <script>
+import SiteNav from './SiteNav.vue'
+
 export default {
+  components: {
+    SiteNav
+  },
+
   props: {
     items: {
       type: Array,
@@ -83,6 +91,7 @@ export default {
 .leftbar {
   width: var(--leftbar-width);
   background: var(--sidebar-bg);
+  padding-right: 30px;
   transition: transform 200ms cubic-bezier(0.2, 1, 0.2, 1);
 
   &.is-hidden {