+++ packages/create-blog/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.7](https://github.com/saberland/saber/compare/create-blog@0.2.6...create-blog@0.2.7) (2019-06-21)
+
+**Note:** Version bump only for package create-blog
+
 ## [0.2.6](https://github.com/egoist/saber/compare/create-blog@0.2.5...create-blog@0.2.6) (2019-05-08)
 
 **Note:** Version bump only for package create-blog