+++ packages/saber-plugin-netlify-redirect/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.1.0](https://github.com/saberland/saber/compare/saber-plugin-netlify-redirect@0.0.6...saber-plugin-netlify-redirect@0.1.0) (2019-07-27)
+
+### Features
+
+- refactor page interface ([#345](https://github.com/saberland/saber/issues/345)) ([629c4a6](https://github.com/saberland/saber/commit/629c4a6))
+
 ## [0.0.6](https://github.com/egoist/saber/compare/saber-plugin-netlify-redirect@0.0.5...saber-plugin-netlify-redirect@0.0.6) (2019-05-21)
 
 ### Bug Fixes