+++ packages/saber-utils/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.2.0](https://github.com/saberland/saber/compare/saber-utils@0.1.5...saber-utils@0.2.0) (2019-07-27)
+
+**Note:** Version bump only for package saber-utils
+
+## [0.1.5](https://github.com/saberland/saber/compare/saber-utils@0.1.4...saber-utils@0.1.5) (2019-06-21)
+
+**Note:** Version bump only for package saber-utils
+
 ## [0.1.4](https://github.com/egoist/saber/compare/saber-utils@0.1.3...saber-utils@0.1.4) (2019-05-11)
 
 **Note:** Version bump only for package saber-utils